Stream JSON-formatted output
This works because only one part of the output is generated, so we can
output a (partial) static blob to start, then an issue at a time, and
finally terminate the object to make it valid.

Because this is a little hinky, the corresponding unit tests have been
adjusted to actually generate (and capture) the entire stdout stream.
We then deserialize (to verify syntax) and compare the resulting dict
with the original input data to confirm there was no content corruption.